+++ src/test/kotlin/no/nav/familie/ba/sak/config/ClientMocks.kt
@@ -8,6 +8,7 @@ import no.nav.familie.ba.sak.common.randomAktørId
 import no.nav.familie.ba.sak.integrasjoner.IntegrasjonClient
 import no.nav.familie.ba.sak.integrasjoner.IntegrasjonOnBehalfClient
 import no.nav.familie.ba.sak.integrasjoner.domene.*
+import no.nav.familie.ba.sak.personopplysninger.domene.AktørId
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Primary
 import org.springframework.context.annotation.Profile
@@ -90,6 +91,43 @@ class ClientMocks {
         every {
             mockIntegrasjonClient.hentPersoninfoFor(eq(søkerFnr[1]))
         } returns Personinfo(fødselsdato = LocalDate.of(1991, 2, 20), kjønn = Kjønn.MANN, navn = "Far Faresen")
+
+        return mockIntegrasjonClient
+    }
+
+    @Profile("test-søk")
+    @Primary
+    @Bean
+    fun mockPDL() : IntegrasjonClient{
+        val mockIntegrasjonClient= mockk<IntegrasjonClient>()
+
+        val farId= "12345678910"
+        val morId= "21345678910"
+        val barnId= "31245678910"
+
+        every {
+            mockIntegrasjonClient.hentPersoninfoFor(farId)
+        } returns Personinfo(fødselsdato = LocalDate.of(1969, 5, 1), kjønn = Kjønn.MANN, navn = "Far Mocksen")
+
+        every {
+            mockIntegrasjonClient.hentPersoninfoFor(morId)
+        } returns Personinfo(fødselsdato = LocalDate.of(1979, 5, 1), kjønn = Kjønn.KVINNE, navn = "Mor Mocksen")
+
+        every {
+            mockIntegrasjonClient.hentPersoninfoFor(barnId)
+        } returns Personinfo(fødselsdato = LocalDate.of(2009, 5, 1), kjønn = Kjønn.MANN, navn = "Barn Mocksen",
+                familierelasjoner = setOf(
+                        Familierelasjoner(Personident(farId), FAMILIERELASJONSROLLE.FAR, "Far Mocksen", LocalDate.of(1969, 5, 1)),
+                        Familierelasjoner(Personident(morId), FAMILIERELASJONSROLLE.MOR, "Mor Mocksen", LocalDate.of(1979, 5, 1))
+                ))
+
+        every {
+            mockIntegrasjonClient.hentAktørId(farId)
+        } returns AktørId(farId)
+
+        every {
+            mockIntegrasjonClient.hentAktørId(barnId)
+        } returns AktørId(barnId)
 
         return mockIntegrasjonClient
     }
